+++ packages/authentication/__mocks__/expo-constants.ts
@@ -1,3 +1,3 @@
-module.exports = require('@baseapp-frontend/test/__mocks__/expo-constants.ts')
+const Constants = require('@baseapp-frontend/test/__mocks__/expo-constants.ts')
 
-export {}
+export default Constants