+++ packages/components/modules/navigations/web/NavMini/__storybook__/stories.tsx
@@ -1,10 +1,10 @@
 import { BaseAppLogoCondensed } from '@baseapp-frontend/design-system/components/web/icons'
+import { defaultTheme } from '@baseapp-frontend/design-system/tests/web/utils'
 
 import { Meta, StoryObj } from '@storybook/react'
 
 import NavMini from '..'
 import { navDataMock } from '../../../../../.storybook/__mocks__/navigation'
-import defaultTheme from '../../../../../.storybook/__mocks__/theme'
 import { NavMiniProps } from '../types'
 
 const meta: Meta<typeof NavMini> = {