+++ packages/components/modules/notifications/web/NotificationsPopover/__tests__/NotificationsPopoverForTesting/index.tsx
@@ -1,8 +1,7 @@
-import withProviders from '../../../../../../__test_utils__/withProviders'
-import { NotificationsPopover, NotificationsPopoverProps } from '../../../index'
+import { withComponentCompleteTestProviders } from '../../../../../tests/web'
+import { NotificationsPopover } from '../../../index'
 
-const NotificationsPopoverForTesting = (props?: Partial<NotificationsPopoverProps>) => {
-  return <NotificationsPopover {...props} />
-}
+const NotificationsPopoverForTestingWithProviders =
+  withComponentCompleteTestProviders(NotificationsPopover)
 
-export default withProviders(NotificationsPopoverForTesting)
+export default NotificationsPopoverForTestingWithProviders