+++ packages/components/modules/tests/web/index.ts
@@ -1,0 +1,3 @@
+export { default as withComponentCompleteTestProviders } from './utils/withComponentCompleteTestProviders'
+export { default as withComponentTestProviders } from './utils/withComponentTestProviders'
+export type * from './utils/withComponentTestProviders/types'