+++ packages/design-system/tests/web/utils/providers/ReactSuspenseWithLoading/types.d.ts
@@ -1,0 +1,6 @@
+import { ComponentType, PropsWithChildren } from 'react'
+
+export interface ReactSuspenseWithLoadingProps extends PropsWithChildren {
+  CustomFallback?: ComponentType
+  disabled?: boolean
+}