+++ packages/design-system/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @baseapp-frontend/design-system
 
+## 1.0.14
+
+### Patch Changes
+
+- Unify test and storybook providers
+- Updated dependencies
+  - @baseapp-frontend/utils@3.1.6
+
 ## 1.0.13
 
 ### Patch Changes